Working on differentiation
The final image is starting to appear in my head.

I'm very excited to use the framework after it's done!

I also think that putting core functions and various gates with losses
is a bad idea. I think it would be better to create another file
specifically made for gates and losses. There are many benefits to
this approach.